Fix Qt 6.9.2 static build linker error with QTimer::singleShot

+++ usagi/src/anidbapi.cpp
@@ -1708,7 +1708,9 @@ void AniDBApi::loadExportQueueState()
 		
 		// Check if export has already been ready (check for existing notification first)
 		// This will be triggered after login when we have a session
-		QTimer::singleShot(5000, this, [this]() { checkForExistingExport(); });
+		// Use static_cast<int> to ensure we call the int overload, not the chrono overload
+		// The chrono overload is not available in static Qt 6.9.2 builds with LLVM MinGW
+		QTimer::singleShot(static_cast<int>(5000), this, [this]() { checkForExistingExport(); });
 	}
 	else
 	{
